closes #45 shard partitioner (#51)
* refs #45 - show the issue

Expose the issue with a test case

Also fix a bug where the testTwoTablesAreOffsetted test reset the ShardedTableMapFile for other tests.  Seemed like a bug anyway

* closes #45 - fix the issue

Add tests and explain how they work.

Also added a strategy to handle missing shards and then make it configurable.
Default to the current implementation.  Still need to figure out how to pass
this config to the MR job.  Welcome better name suggestions for the property.

* closes #45 - address PR feedback