Added PasswordConverter to allow passwords from environment
Created PasswordConverter class to allow users to specify passwords
using an env: prefix.  This will indicate that an ENVIRONMENT variable
should be used to obtain the password.

This is useful in order to prevent passwords from showing up in the process list

Updated some core classes to make use of this convention.