Added updates to help validate shard splits (#111)
* Added updates to help validate shard splits

Updated ShardedTableMapFile.setupFile method to take an additional boolean
argument specifying whether validation of the generated shard to tablet mappings
should be validated for a given date range

Validation is turned off by default in IngestJob unless the:
- shardedMap.validation.enabled
property is set to true

Default is to validate the last two days of shard table splits.  The range can be
overridden using the:
- shards.balanced.days.to.verify
property, where 1 would mean only validate today, 2 yeterday and today, etc.

Validation is dependent on the ShardIdFactory class to inform us of the number
of shards that should exist for a given day.  Any deviation from this number,
when validation is turned on, will result in an IllegalStateException

Validation that the shards are balanced, i.e. have been spread out to unique
tservers via custom balancer, is also performed when validation is on. An
unbalanced mapping of shard splits will also result in an IllegalStateException.

Removed boolean param since it is being passed in the conf object
which is accessible in ShardedTableMapFile, default behavior
is to not do validation

Also set log level to error for shard creation failure log messages