Updated FileRuleWatcher load parent from child dir
Modified FileRuleWatcher to use the child config dir as the root
when loading a parent config.  This should simpliy the specification
of the parent path to just the relative name.

Added some more tests and fixed the merging of matchpattern for the
case that a match pattern is not specified in the child config.