File rule watcher load path fix (#181)
* Fix for FileRuleWatcher where path to file in hdfs not loaded correctly

Previously used this.getClass.getResource() to look up file.  This will
not necessarily resolve files with hdfs:// prefixes, which results in an
unintended IllegalArgumentException

* Fix for FileRuleWatcher to correct unit test failures

* Updated FileRuleWatcher load parent from child dir

Modified FileRuleWatcher to use the child config dir as the root
when loading a parent config.  This should simpliy the specification
of the parent path to just the relative name.

Added some more tests and fixed the merging of matchpattern for the
case that a match pattern is not specified in the child config.