Modified Shard Balance check to allow multiple per tserver (#323)
* Modified Shard Balance check to allow multiple per tserver

New shardedMap.max.tablets.per.shard config property will allow a maximum
number of shards per tablet.  This is useful for the case that there
are more shards then there are tservers on the system. We still want to be
balanced as much as possible, but some can be on more then one up to max.

* Updated property and variable name per feedback