Remove hack that replaced authentication cache. re/fixes #380
Wildlfy maintains an authentication cache, and previously due
to the bug WFLY-3858, the authentication cache could not be
configured using infinispan where we could set the lifespan of
entries. Therefore, a hack was introduced to replace the
authentication cache with a Guava cache. However, with the
upgrade with the upgrade to Wildfly 16, the original bug is
now fixed and we can once again use infinispan to configure the
authentication cache lifespan.