Move edge model configuration to a separate module.
When the edge dictionary was split out to a separate service
the edge-dictionary-core module was repurposed to use some
wildfly CDI-spring bridge code. This module is now included
in extended clients. Previously, this module was not included
in extended clients. The problem is that in addition to the
edge dictionary, the module also contained the edge model
configuration. Extended clients would then get two edge model
configurations and it would be at the whim of the classpath
as to which got loaded. Therefore, move the edge model config
to its own module that can be included or excluded as desired.