Fix github actions issue--move docker root.
The quickstart tests were hanging since the
volume on which the docker data is stored
started crossing 90% full, which triggered
yarn to mark the disk as unhealthy and then
the MapReduce jobs would never run.

We re-configure docker to move the data root
to the 14GB ssd drive instead, and this clears
enough space for the job to run.