Add support for multiple ActiveQueryLog instances
Currently an ActiveQueryLog singleton is used for logging events from
all active queries. It can be difficult to easily search the logs for
queries that are running on specific tables.

Add support to be able to retrieve a new ActiveQueryLog instance
associated with a configurable name, and include this name in log
messages generated by the ActiveQueryLog. Keep support for a default
ActiveQueryLog instance that can be shared by queries when additional
search details is not needed.

In addition, add the ability to specifiy a source for the active query
log name for ShardQueryLogic.

Fixes #817