Preserve previous delete behavior
Add a timestampsIgnored option, with a default of true. When this option
is enabled, then the aggregrator will retain its previous behavior of
always giving precedence to a deleted UID if both an add and a delete of
the same UID are aggregated together. When the timestampsIgnored option
is disabled, then the order of iteration (which is in timestamp order
of newer to older timestamps) is used to give precedence. A remove
occurring after an add will remove the UID and an add occurring after a
remove will re-add the UID.