Fix reparented query property marker bug
Currently when the sources are retrieved for a query property marker via
the QueryPropertyMarkerVisitor, if the marker has multiple sources, they
are returned as children of a new AND node and are subsequently
reparented.

Modify the source retrieval code so that the source nodes are not
reparented as children of the new AND node. Additionally, create
visitors to fix query property marker nodes with multiple sources and
to validate query trees to ensure that they do not have multiple
sources.

Fixes #603