Fix occurrence query function
The occurrence function currently does not count occurrences correctly.
Add the ability to track when a field value originated from the index
vs. the event via its Attribute. Modify the occurrence function to only
return the total number of values from the event, unless there are none.
In this case, return the total number of values from the index.

Fixes #1108