Fix occurrence query function (#1371)
* Fix occurrence query function

The occurrence function currently does not count occurrences correctly.
Add the ability to track when a field value originated from the index
vs. the event via its Attribute. Modify the occurrence function to only
return the total number of values from the event, unless there are none.
In this case, return the total number of values from the index.

Fixes #1108

* Initialize Attributes.fromIndex as true

* Make DocumentData.fromIndex non-mutable

* Make ValueToAttributes.fromIndex final

* Add support for non-ValueTuple objects

* Remove unused import

Co-authored-by: Ivan Bella <ivan@bella.name>
Co-authored-by: hgklohr <hgklohr@gmail.com>