Fix CustomAnalyzerQueryNodeProcessor to avoid dropping tokens when generating a query that contains synonyms. (#1412)
* test case for incomplete query tokenization issue

* Bugfixes for CustomAnalyzerQueryNodeProcessor

* Fixed issue where tokens would be ignored if the first token matched
the input query term. Some Analyzers produce term variants with the
initial term being identical to the input term. This will slightly
change behavior when dealing with slop, but this will make little
practical difference because we avoid slop when tokenizing because it
does not peserve word order (e.g., the slop handling code should be
deprecated).
* Fixed issue where we were checking for the presence of the
NODE_PROCESSED tag instead of actually checking that the value equals
Boolean.TRUE.
* Cleaned up some warninigs

Co-authored-by: hgklohr <hgklohr@gmail.com>