updated code to support direct proxied server roles (#1981)
Added additional verification for terminal proxy server. Now the terminal server (the server issuing the call) must pass a positive check of either having AuthorizedQueryServer or AuthorizedServer. An additional role, AuthorizedProxiedServer, allows a server to be in XPE chain, but it must be proxied by another server. It can not be the terminal server.