Fix #GROUPBY duplicate counts with query model
When query model mapping is applied to a document before grouping and
aggregating, it is possible for multiple fields to be mapped to the same
root model mapping with the same value, resulting in duplicate
field-value pairings that were incorrectly included in grouping and
aggregation.

Ensure that duplicate field-value pairings are disregarded. Add tests to
verify this and update documentation to include details about how model
mapping affects the #GROUPBY function.

Fixes #2360