Attempt to work through in-memory-accumulo issues
Tried to build against accumulo 4.0.0-SNAPSHOT. Found errors with
InMemoryAccumulo. Attempted to resolve them by adding a instanceID
method to InMemoryAccumulo and swapping it for all missing instances of
InMemoryInstance.

Ran into errors with not having push access to submodule
datawave-in-memory-accumulo.

Think it might be less effort to create a cachingScanner vs continue
using the in-memory accumulo instance.