Add MAX_UNIQUE_COUNT functionality
Add the ability to specify a max count on results from the unique
operation such that only the unique results that occurred up to the
specified max count are returned.

This max count can be specified either via the function
MAX_UNIQUE_COUNT or via the query parameter max.unique.count.

Fixes #2635