Initial commit for experimental scheduler that splits query execution into two phases: field index scan and document scan.
The executors that handle field index scans or document scans can be
scaled independently of one another. The scheduler can submit more tasks
than there are threads to improve throughput in the case of many short
lived tasks.

The executors are effectively a cached threadpool, so unused threads
will be terminated after a waiting period.

Iterator and timing stats are collected and persisted for the field
index scan and logged when the scheduler is closed.

TLDEventQuery and FieldIndexQuery are explicitly not supported

All executable markers are handled (value, range, list)

Add min/avg/max and p50,p95,p99 stats for document candidate and retrieval