Returned detailed message when query validation rules not configured
Currently, when query validation rules are not configured for a
ShardQueryLogic bean, the resulting exception and message is identical
to the default UnsupportedOperationException thrown by the QueryLogic
interface. This is misleading, since the operation is supported, but
additional configuration is required.

Update the ShardQueryLogic to instead throw a BadRequestQueryException
with the error code 412-19 and a message indicating that query
validation rules need to be configured for the underlying logic before
validation can proceed.

Additionally, add the option -VQ | --validateQuery to the quickstart
datawaveQuery function to make the command post a query to the validate
endpoint.