Prevent Docker Quickstart from using incorrect MAVEN_HOME path (#3065)
The MAVEN_HOME ENV variable set in the Dockerfile is not used in the
COPY / ADD lines. Instead, the host's MAVEN_HOME path is used and will
lead to a "file not found" error if it is anything other than
`/usr/share/maven`.

This resolves this issue by hardcoding the paths instead.

Fixes #3064