Allow suspend / resume of IvaratorRunnable after it was stuck during suspend; reuse the IvaratorRunnables that suspended successfully (#3223)
Allow suspend / resume of IvaratorRunnable after it was stuck during
suspend; reuse the IvaratorRunnables that suspended successfully

When the DatawaveFieldIndexCachingIterator is yielding, it suspends any
IvaratorRunables that are still executing via a call through the
IteratorThreadPoolManager that sets suspendRequested and then waits for
up to 60 seconds for the thread to start (if it hasn't already) and then
up to 60 seconds for the thread to check suspendRequested and suspend.
We have seen cases where the 60 seconds to complete elapse, the
DatawaveFieldIndexCachingIterator finishes the yield, and then the
IvaratorRunnable gets an exception when trying to use the Accumulo
source that the Tablet has reclaimed. This exception seems to happen
very quickly after control is returned to Tablet.

If the IvaratorRunnable already has a next key, then it still might be
able to move to a SUSPEND status before the
DatawaveFieldIndexCachingIterator is re-seeked after the yield. The set
and get of status are synchronized, so this should be safe.

Also related is that the DatawaveFieldIndexCachingIterator is currently
re-starting all IvaratorRunnables if there is one that is not CREATED or
SUSPENDED or COMPLETED. So if one IvaratorRunnable fails to suspend,
then re re-run all of them. This is overly conservative. We should be
able to remove the FAILED IvaratorRunnable / IvaratorFuture and then
re-run just that one while resuming all others.

This change also moves the duration and TimeUnit from IvaratorRunnable
to one place -- the call to IteratorThreadPoolManager.suspendIvarator
and uses that duration instead of the 60 + 60 second waits that were
previously there. This should make the code more readable and the wait
durations less opaque.